pyproject.toml
[tool.poetry]
name = "xfyun-services"
version = "0.1.0"
description = ""
authors = ["fruitbars <huiyanghu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
websocket = "^0.2.1"
websocket-client = "^1.8.0"
websockets = "^14.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
